Extracted code to perform detection in its own Python package
The code in spack.cmd.external related to package detection
has been moved to the spack.detection package. Added
docstrings to functions.
